requirements.txt
python-telegram-bot>=20.0
httpx>=0.24.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
